Significant refactor of backend, alien domain countermeasure in place…
files saving with file extension and working throughout however seems to be temperamental as encoded files are the same size as originals, also when file extension is invovled in encoding, a steg.py exception is raised. We need to test the steg.py thoroughly to identify the cause of the issue. REMOVAL OF PRINT STATEMENTS